+++ sdk/core/azure-core/inc/azure/core/operation.hpp
@@ -31,6 +31,12 @@ namespace Azure { namespace Core {
     virtual Response<T> PollUntilDoneInternal(std::chrono::milliseconds period, Context& context)
         = 0;
 
+    [[deprecated("Do not override and do not use.")]] virtual Azure::Core::Http::RawResponse const&
+    GetRawResponseInternal() const
+    {
+      return *m_rawResponse;
+    }
+
   protected:
     /** @brief the underlying raw response for this operation. */
     std::unique_ptr<Azure::Core::Http::RawResponse> m_rawResponse = nullptr;